Split client msgs
