+++ modules/src/ics02_client/handler/update_client.rs
@@ -6,7 +6,7 @@ use crate::ics02_client::context::ClientReader;
 use crate::ics02_client::error::{Error, Kind};
 use crate::ics02_client::handler::{ClientEvent, ClientResult};
 
-use crate::ics02_client::msgs::MsgUpdateAnyClient;
+use crate::ics02_client::msgs::update_client::MsgUpdateAnyClient;
 use crate::ics24_host::identifier::ClientId;
 
 /// The result following the successful processing of a `MsgUpdateAnyClient` message. Preferably
@@ -70,7 +70,8 @@ mod tests {
     use crate::ics02_client::handler::ClientResult::{Create, Update};
     use crate::ics02_client::handler::{dispatch, ClientEvent};
     use crate::ics02_client::header::Header;
-    use crate::ics02_client::msgs::{ClientMsg, MsgUpdateAnyClient};
+    use crate::ics02_client::msgs::update_client::MsgUpdateAnyClient;
+    use crate::ics02_client::msgs::ClientMsg;
     use crate::ics24_host::identifier::ClientId;
     use crate::mock_client::header::MockHeader;
     use crate::mock_client::state::MockClientState;
