+++ relayer/src/chain/cosmos.rs
@@ -26,7 +26,8 @@ use ibc_proto::cosmos::base::v1beta1::Coin;
 use ibc_proto::cosmos::tx::v1beta1::mode_info::{Single, Sum};
 
 use ibc::ics02_client::client_def::{AnyClientState, AnyConsensusState, AnyHeader};
-use ibc::ics02_client::msgs::{MsgCreateAnyClient, MsgUpdateAnyClient};
+use ibc::ics02_client::msgs::create_client::MsgCreateAnyClient;
+use ibc::ics02_client::msgs::update_client::MsgUpdateAnyClient;
 use ibc::ics03_connection::connection::{ConnectionEnd, Counterparty};
 use ibc::ics07_tendermint::client_state::ClientState;
 use ibc::ics07_tendermint::consensus_state::ConsensusState as TendermintConsensusState;
@@ -127,7 +128,11 @@ impl CosmosSDKChain {
             signed_header,
             validator_set,
             next_validator_set,
-            self.config().peers.clone().ok_or_else(|| Kind::Config.context("no peers configured".to_string()))?.primary,
+            self.config()
+                .peers
+                .clone()
+                .ok_or_else(|| Kind::Config.context("no peers configured".to_string()))?
+                .primary,
         );
 
         Ok(light_block)
